Need separate descriptor tables for tracking open/close vs map/munmap.
Before, we had only a single descriptor table that tracked open/close.
The problem is that the application sequence "open/map/close ... munmap"
is under-specified: either on "close" we destory the memfile, or
we keep the memfile open, but have no way to destroy it on
munmap.  Now, the open returns an fd and map returns a mapfd,
and, for each client, the server maintains two tables: one for
each descriptor type.